refactor(budget): extract the budget model into a leaf crate
The whole budget model — reading Fava's `custom "budget"` directives, the
calendar interval arithmetic, supersession, and the per-day accrual — sat
private inside the `rledger` binary's report module. No other consumer could
reach it: rustfava, the FFI component and the query engine would each have
had to re-derive per-(account,currency) supersession, calendar anchoring and
pro-rating against their own types. Those rules are subtle enough that two
implementations would certainly drift; the leap-February denominator and the
mid-interval supersession split are exactly the kind of thing that diverges
quietly and disagrees only on someone's real ledger.

`rustledger-returns` exists for the same reason and set the precedent, so
this follows it: `rustledger-budget` owns the model and does no ledger
loading, no rendering and no I/O. The report module keeps only what is
genuinely report-specific — pairing budgets with actual spending, the
`--children` and `--account` selection rules, totals and rendering — and
drops from ~770 lines to ~560.

The move also collapses a duplication the architecture review flagged: the
"next declaration after this date" scan was written twice inside `accrue`
with only the surrounding control flow differing, and `in_force` rescanned
the whole flat list on every iteration. Both are now methods on a `Budgets`
index, so the supersession predicate exists once. A change to it can no
longer make the "skip ahead" branch and the "segment end" branch disagree
about the rate for a segment.

Behavior is unchanged: the same 10 unit tests (9 now in the crate, 1 on the
report's own row arithmetic) and 20 integration tests pass untouched.

Adds the crate to the release-publish crate list. Omitting a new crate there
is what broke the v0.16.0 crates.io publish, and the list is order-sensitive.
